# Remove unnecessary looping in field data cache clear