### Description
Previously, clearing a shard from the field data cache (which happens on shard closure like in a B/G, or during manual cache clear) unnecessarily looped through each of the N keys in the whole cache, once per field that uses field data on that index, which can potentially be very large. This takes O(FN) time for F the number of fields on that index. For real user scenarios this could get as long as minutes which caused node drops. 

This PR moves around the cleanup logic a bit, so that the IndicesFieldDataCache itself is the one that iterates through its keys when a cleanup is requested. This takes us from O(FN) to O(N) time to clear a shard.

In the next PR, the iteration itself will be moved to a scheduled cleaner thread, as in the request cache, and the clear() methods on IndicesFieldDataCache will mark those indices/fields for cleanup in that thread, rather than iterating through them then and there. I purposefully split these, because in the case of a whole-cache clean via API call, the existing code in the request cache actually has a similar issue, and I wanted to raise an issue to get feedback on that and then apply the same solution to both the FDC and the RC. 

A future PR will also remove the synchronized blocks in IndexFieldDataService. This depends on the scheduled cleaner thread to work without iteration issues on the cache keys iterator, so it's also separate.

Benchmark results: 
On a single-node OS cluster running on a c5.2xl EC2 instance, I set up indices with 999 field data-enabled fields each. 999 is the max number of fields per index without adjusting settings; it's reasonable to pick such a high value because the users who saw issues were actually using ~20k fields each. Then I searched each field for each index to populate the field data cache with many keys. Finally, I cleared one index (which has 1 shard) using the cache clear API and timed it on the client side (end-to-end). Note this hits the same code path as when the shard closes in a B/G. Scripts to do this are [here](https://github.com/peteralfonsi/misc-scripts/tree/main/cache-clear-timing/fielddata). 

|Number of indices | Baseline clear time (sec) | Contender clear time (sec) | 
|--|--|--|
|100|59.09|0.0993|
|1|0.0153|0.0186|

I purposefully did not time clearing the whole cache at once as this will be changed to the scheduled thread in the next PR. 

We see for 100 indices the speedup is very large. This is because N is very large so checking all N keys 999 times is slow. There isn't a speedup when there's just 1 index; I think the 0.01 sec is mostly overhead of OS processing the request rather than actual cache clear time. 


### Related Issues
Part of https://github.com/opensearch-project/OpenSearch/issues/13862

### Check List
- [x] Functionality includes testing.
- [N/A] API changes companion pull request [created](https://github.com/opensearch-project/opensearch-api-specification/blob/main/DEVELOPER_GUIDE.md), if applicable.
- [N/A] Public documentation issue/PR [created](https://github.com/opensearch-project/documentation-website/issues/new/choose), if applicable.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).